split code up into two pages
